requirements.txt
customtkinter==5.0.3
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.3.0
